Fix int-double bug in SubmissionDto x2, change scale to 2

+++ src/main/java/com/a6raywa1cher/coursejournalbackend/rest/dto/BatchSetForStudentAndCourseSubmissionRestDto.java
@@ -31,6 +31,6 @@ public class BatchSetForStudentAndCourseSubmissionRestDto {
         private ZonedDateTime submittedAt;
 
         @NotNull
-        private Integer additionalScore;
+        private Double additionalScore;
     }
 }
